+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to
 
 ## Unreleased
 
+### Fixed
+
+- Fix step failure on `MISSING_KEY_ERROR` when group not found building
+  relationship to application. The integration will now log a warn message and
+  continue ingestion of other relationships.
+
+- Fix step failure on `MISSING_KEY_ERROR` when user not found building
+  relationship to application. The integration will now log a warn message and
+  continue ingestion of other relationships.
+
 ## 2.1.0 - 2021-07-22
 
 ### Added